Refactor group detail constituent architecture
