+++ .beads/interactions.jsonl
@@ -98,3 +98,5 @@
 {"id":"int-7023f542","kind":"field_change","created_at":"2026-06-26T23:50:29.445364Z","actor":"David Ten","issue_id":"stockscreenclaude-0mq","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Completed: group detail constituents now load from the requested market scan/feature run and group charts render two per row."}}
 {"id":"int-405fbd31","kind":"field_change","created_at":"2026-06-27T00:03:59.634451Z","actor":"David Ten","issue_id":"stockscreenclaude-lvg","extra":{"field":"status","new_value":"in_progress","old_value":"open"}}
 {"id":"int-9704dbdf","kind":"field_change","created_at":"2026-06-27T00:15:17.087419Z","actor":"David Ten","issue_id":"stockscreenclaude-lvg","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Completed thermo review cleanup for group detail constituents"}}
+{"id":"int-ae392f65","kind":"field_change","created_at":"2026-06-27T00:26:36.158967Z","actor":"David Ten","issue_id":"stockscreenclaude-k82","extra":{"field":"status","new_value":"in_progress","old_value":"open"}}
+{"id":"int-6ae6efc4","kind":"field_change","created_at":"2026-06-27T00:38:38.446054Z","actor":"David Ten","issue_id":"stockscreenclaude-k82","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Fixed group detail review architecture findings"}}
