+++ .beads/issues.jsonl
@@ -1,3 +1,4 @@
+{"id":"stockscreenclaude-k82","title":"Fix group detail review architecture findings","status":"closed","priority":1,"issue_type":"task","owner":"davidten7@gmail.com","created_at":"2026-06-27T00:26:13Z","created_by":"David Ten","updated_at":"2026-06-27T00:38:38Z","started_at":"2026-06-27T00:26:36Z","closed_at":"2026-06-27T00:38:38Z","close_reason":"Fixed group detail review architecture findings","dependency_count":0,"dependent_count":0,"comment_count":0}
 {"id":"stockscreenclaude-lvg","title":"Address group detail thermo review findings","status":"closed","priority":1,"issue_type":"task","owner":"davidten7@gmail.com","created_at":"2026-06-27T00:03:49Z","created_by":"David Ten","updated_at":"2026-06-27T00:15:17Z","started_at":"2026-06-27T00:04:00Z","closed_at":"2026-06-27T00:15:17Z","close_reason":"Completed thermo review cleanup for group detail constituents","dependency_count":0,"dependent_count":0,"comment_count":0}
 {"id":"stockscreenclaude-0mq","title":"Fix group detail constituents and two-column charts","status":"closed","priority":1,"issue_type":"bug","owner":"davidten7@gmail.com","created_at":"2026-06-26T23:43:12Z","created_by":"David Ten","updated_at":"2026-06-26T23:50:29Z","started_at":"2026-06-26T23:43:28Z","closed_at":"2026-06-26T23:50:29Z","close_reason":"Completed: group detail constituents now load from the requested market scan/feature run and group charts render two per row.","dependency_count":0,"dependent_count":0,"comment_count":0}
 {"id":"stockscreenclaude-05y","title":"Fix daily delta refresh guard denominator for GitHub-seeded top-ups","status":"closed","priority":1,"issue_type":"bug","owner":"davidten7@gmail.com","created_at":"2026-06-24T06:12:18Z","created_by":"David Ten","updated_at":"2026-06-24T07:10:24Z","started_at":"2026-06-24T06:12:25Z","closed_at":"2026-06-24T07:10:24Z","close_reason":"Addressed follow-up accounting review findings","dependency_count":0,"dependent_count":0,"comment_count":0}
