+++ backend/app/services/static_site_export_service.py
@@ -29,7 +29,6 @@ from app.domain.scanning.default_filters import (
 from app.infra.db.models.feature_store import FeatureRun, FeatureRunPointer
 from app.infra.db.repositories.feature_store_repo import SqlFeatureStoreRepository
 from app.schemas.groups import (
-    ConstituentStock,
     GroupDetailResponse,
     GroupRankResponse,
     GroupRankingsResponse,
@@ -38,6 +37,10 @@ from app.schemas.groups import (
 )
 from app.schemas.scanning import FilterOptionsResponse, ScanResultItem
 from app.services.breadth_attribution_service import BreadthAttributionService
+from app.services.group_detail_payloads import (
+    constituent_stock_payloads_from_group_rows,
+    scan_result_item_to_group_row,
+)
 from app.services.market_exposure_service import build_exposure_payload
 from app.services.preset_screens import (
     PRESET_SCREENS,
@@ -1344,7 +1347,7 @@ class StaticSiteExportService:
         *,
         ranking_date: date,
     ) -> list[dict[str, Any]]:
-        normalized_rows = [self._extract_group_row_payload(row) for row in rows]
+        normalized_rows = [scan_result_item_to_group_row(row) for row in rows]
         return self._compute_group_rankings_from_serialized_rows(
             normalized_rows,
             ranking_date=ranking_date,
@@ -1428,33 +1431,6 @@ class StaticSiteExportService:
         for index, row in enumerate(rankings, start=1):
             row["rank"] = index
         return rankings
-
-    @staticmethod
-    def _extract_group_row_payload(row: Any) -> dict[str, Any]:
-        extended = getattr(row, "extended_fields", {}) or {}
-        return {
-            "symbol": getattr(row, "symbol", None),
-            "company_name": extended.get("company_name"),
-            "composite_score": getattr(row, "composite_score", None),
-            "current_price": getattr(row, "current_price", None),
-            "rs_rating": extended.get("rs_rating"),
-            "rs_rating_1m": extended.get("rs_rating_1m"),
-            "rs_rating_3m": extended.get("rs_rating_3m"),
-            "rs_rating_12m": extended.get("rs_rating_12m"),
-            "eps_growth_qq": extended.get("eps_growth_qq"),
-            "eps_growth_yy": extended.get("eps_growth_yy"),
-            "sales_growth_qq": extended.get("sales_growth_qq"),
-            "sales_growth_yy": extended.get("sales_growth_yy"),
-            "stage": extended.get("stage"),
-            "market_cap": extended.get("market_cap"),
-            "market_cap_usd": extended.get("market_cap_usd"),
-            "ibd_industry_group": extended.get("ibd_industry_group"),
-            "price_sparkline_data": extended.get("price_sparkline_data"),
-            "price_trend": extended.get("price_trend"),
-            "price_change_1d": extended.get("price_change_1d"),
-            "rs_sparkline_data": extended.get("rs_sparkline_data"),
-            "rs_trend": extended.get("rs_trend"),
-        }
 
     @staticmethod
     def _group_rank_map(rankings: list[dict[str, Any]]) -> dict[str, dict[str, Any]]:
@@ -1566,37 +1542,9 @@ class StaticSiteExportService:
                     ).model_dump(mode="json")
                 )
 
-            stocks = sorted(
-                current_rows_by_group.get(group_name, []),
-                key=lambda row: (
-                    row.get("rs_rating") if row.get("rs_rating") is not None else float("-inf"),
-                    row.get("composite_score") if row.get("composite_score") is not None else float("-inf"),
-                ),
-                reverse=True,
+            stock_payload = constituent_stock_payloads_from_group_rows(
+                current_rows_by_group.get(group_name, [])
             )
-            stock_payload = [
-                ConstituentStock(
-                    symbol=row["symbol"],
-                    company_name=row.get("company_name"),
-                    price=row.get("current_price"),
-                    rs_rating=row.get("rs_rating"),
-                    rs_rating_1m=row.get("rs_rating_1m"),
-                    rs_rating_3m=row.get("rs_rating_3m"),
-                    rs_rating_12m=row.get("rs_rating_12m"),
-                    eps_growth_qq=row.get("eps_growth_qq"),
-                    eps_growth_yy=row.get("eps_growth_yy"),
-                    sales_growth_qq=row.get("sales_growth_qq"),
-                    sales_growth_yy=row.get("sales_growth_yy"),
-                    composite_score=row.get("composite_score"),
-                    stage=row.get("stage"),
-                    price_sparkline_data=row.get("price_sparkline_data"),
-                    price_trend=row.get("price_trend"),
-                    price_change_1d=row.get("price_change_1d"),
-                    rs_sparkline_data=row.get("rs_sparkline_data"),
-                    rs_trend=row.get("rs_trend"),
-                ).model_dump(mode="json")
-                for row in stocks
-            ]
 
             details[group_name] = GroupDetailResponse(
                 industry_group=group_name,
